+++ bins/revme/src/cmd/eofvalidation.rs
@@ -1,6 +1,6 @@
 mod test_suite;
 
-pub use test_suite::{PragueTestResult, TestResult, TestSuite, TestUnit, TestVector};
+pub use test_suite::{TestResult, TestSuite, TestUnit, TestVector};
 
 use crate::{cmd::Error, dir_utils::find_all_json_tests};
 use revm::interpreter::analysis::{validate_raw_eof_inner, CodeType, EofError};
@@ -76,18 +76,25 @@ pub fn run_test(path: &Path) -> Result<(), Error> {
                     continue;
                 }
                 test_sum += 1;
-                let kind = if test_vector.container_kind.is_some() {
-                    Some(CodeType::ReturnContract)
-                } else {
-                    Some(CodeType::ReturnOrStop)
+                let kind = match test_vector.container_kind.as_deref() {
+                    Some("RUNTIME") => Some(CodeType::ReturnOrStop),
+                    Some("INITCODE") => Some(CodeType::ReturnContract),
+                    None => Some(CodeType::ReturnOrStop),
+                    _ => return Err(Error::Custom("Invalid container kind")),
+                };
+                // In future this can be generalized to cover multiple forks, Not just Osaka.
+                let Some(test_result) = test_vector.results.get("Osaka") else {
+                    // if test does not have a result that we can compare to, we skip it
+                    println!("Test without result: {} - {}", name, vector_name);
+                    continue;
                 };
                 let res = validate_raw_eof_inner(test_vector.code.clone(), kind);
-                if res.is_ok() != test_vector.results.prague.result {
+                if test_result.result != res.is_ok() {
                     println!(
                         "\nTest failed: {} - {}\nresult:{:?}\nrevm err_result:{:#?}\nbytes:{:?}\n",
                         name,
                         vector_name,
-                        test_vector.results.prague,
+                        test_result.result,
                         res.as_ref().err(),
                         test_vector.code
                     );