+++ crates/revm/src/handler/mainnet/pre_execution.rs
@@ -7,8 +7,8 @@ use crate::{
     primitives::{
         db::Database,
         eip7702, Account, Bytecode, EVMError, Env, Spec,
-        SpecId::{CANCUN, PRAGUE},
-        TxKind, BLOCKHASH_STORAGE_ADDRESS, KECCAK_EMPTY, U256,
+        SpecId::{CANCUN, PRAGUE, SHANGHAI},
+        TxKind, KECCAK_EMPTY, U256,
     },
     Context, ContextPrecompiles,
 };
@@ -38,16 +38,6 @@ pub fn load_accounts<SPEC: Spec, EXT, DB: Database>(
     //         .warm_preloaded_addresses
     //         .insert(coinbase);
     // }
-
-    // Load blockhash storage address
-    // EIP-2935: Serve historical block hashes from state
-    if SPEC::enabled(PRAGUE) {
-        context
-            .evm
-            .journaled_state
-            .warm_preloaded_addresses
-            .insert(BLOCKHASH_STORAGE_ADDRESS);
-    }
 
     // Load access list
     context.evm.load_access_list()?;